Lazy SMP 1 thread +8 Elo Lazy SMP 2 threads +70 Elo vs 1 thread Lazy SMP 4 threads +110 Elo vs 1 thread Lazy SMP 8 threads +169 Elo vs 1 thread
Score of Pedantic Dev vs Pedantic Base: 1684 - 1546 - 3170  [0.511] 6400
...      Pedantic Dev playing White: 978 - 636 - 1587  [0.553] 3201
...      Pedantic Dev playing Black: 706 - 910 - 1583  [0.468] 3199
...      White vs Black: 1888 - 1342 - 3170  [0.543] 6400
Elo difference: 7.5 +/- 6.0, LOS: 99.2 %, DrawRatio: 49.5 %
SPRT: llr 2.2 (100.3%), lbound -2.2, ubound 2.2 - H1 was accepted

Score of Pedantic Dev vs Pedantic Base: 257 - 121 - 308  [0.599] 686
...      Pedantic Dev playing White: 143 - 52 - 148  [0.633] 343
...      Pedantic Dev playing Black: 114 - 69 - 160  [0.566] 343
...      White vs Black: 212 - 166 - 308  [0.534] 686
Elo difference: 69.8 +/- 19.4, LOS: 100.0 %, DrawRatio: 44.9 %
SPRT: llr 2.21 (100.5%), lbound -2.2, ubound 2.2 - H1 was accepted

Score of Pedantic Dev vs Pedantic Base: 149 - 37 - 180  [0.653] 366
...      Pedantic Dev playing White: 86 - 14 - 83  [0.697] 183
...      Pedantic Dev playing Black: 63 - 23 - 97  [0.609] 183
...      White vs Black: 109 - 77 - 180  [0.544] 366
Elo difference: 109.8 +/- 25.3, LOS: 100.0 %, DrawRatio: 49.2 %
SPRT: llr 2.21 (100.4%), lbound -2.2, ubound 2.2 - H1 was accepted

Score of Pedantic Dev vs Pedantic Base: 130 - 17 - 103  [0.726] 250
...      Pedantic Dev playing White: 75 - 7 - 43  [0.772] 125
...      Pedantic Dev playing Black: 55 - 10 - 60  [0.680] 125
...      White vs Black: 85 - 62 - 103  [0.546] 250
Elo difference: 169.3 +/- 33.6, LOS: 100.0 %, DrawRatio: 41.2 %
SPRT: llr 2.2 (100.1%), lbound -2.2, ubound 2.2 - H1 was accepted
