+++ Pedantic.Chess/Uci.cs
@@ -1,159 +1,150 @@
-﻿// ***********************************************************************
-// Assembly         : Pedantic.Chess
-// Author           : JoAnn D. Peeler
-// Created          : 01-17-2023
-//
-// Last Modified By : JoAnn D. Peeler
-// Last Modified On : 03-27-2023
-// ***********************************************************************
-// <copyright file="Uci.cs" company="Pedantic.Chess">
-//     Copyright (c) . All rights reserved.
-// </copyright>
-// <summary>
-//     A class that encapsulates UCI responses from the engine.
-// </summary>
-// ***********************************************************************
-using Pedantic.Tablebase;
+﻿using System.Text;
 using Pedantic.Utilities;
-using System.Text;
 
 namespace Pedantic.Chess
 {
-    public static class Uci
+    public class Uci
     {
-        private static bool disableOutput;
-        private static readonly object lockObject = new();
-
-        public static bool DisableOutput
+        static Uci()
         {
-            get
-            {
-                lock (lockObject)
-                {
-                    return disableOutput;
-                }
-            }
-            set
-            {
-                lock (lockObject)
-                {
-                    disableOutput = value;
-                }
-            }
+            defaultUci = new Uci(true, false);
         }
 
-        public static void Log(string message)
+        public Uci(bool enable = true, bool debug = false)
         {
-            Console.Out.WriteLineAsync($@"info string {message}");
+            this.enable = enable;
+            this.debug = debug;
         }
 
-        public static void Debug(string message)
+        public void Log(string message)
         {
-            if (Engine.Debug)
+            if (!enable)
+            {
+                return;
+            }
+            Console.Out.WriteLineAsync($"info string {message}");
+        }
+
+        public void Debug(string message)
+        {
+            if (!enable)
+            {
+                return;
+            }
+            if (debug)
             {
                 Console.Out.WriteLineAsync($"info string {message}");
             }
         }
 
-        public static void Info(int depth, int seldepth, int score, long nodes, long timeMs, ulong[] pv, 
+        public void Info(int depth, int seldepth, int score, long nodes, long timeMs, ulong[] pv,
             int hashfull, long tbHits, TtFlag flag = TtFlag.Exact)
         {
+            if (!enable)
+            {
+                return;
+            }
             StringBuilder sb = new();
             int nps = (int)(nodes * 1000 / Math.Max(1, timeMs));
-            sb.Append(@$"info depth {depth} seldepth {seldepth} score cp {score}");
+            sb.Append($"info depth {depth} seldepth {seldepth} score cp {score}");
             if (flag == TtFlag.UpperBound)
             {
                 sb.Append(" upperbound");
             }
-            else if (flag == TtFlag.LowerBound)
+            else if (flag == TtFlag.LowerBound) 
             {
                 sb.Append(" lowerbound");
             }
-            sb.Append(@$" nodes {nodes} hashfull {hashfull} nps {nps} time {timeMs} ");
+            sb.Append($" nodes {nodes} hashfull {hashfull} nps {nps} time {timeMs}");
             if (tbHits > 0)
             {
-                sb.Append(@$"tbhits {tbHits} ");
+                sb.Append($" tbhits {tbHits}");
             }
-
-            sb.Append("pv");
+            sb.Append(" pv");
             for (int n = 0; n < pv.Length; n++)
             {
-                sb.Append($@" {Move.ToString(pv[n])}");
+                sb.Append($" {Move.ToString(pv[n])}");
             }
-
             string output = sb.ToString();
-
-            if (!DisableOutput)
-            {
-                Console.Out.WriteLineAsync(output);
-            }
+            Console.Out.WriteLineAsync(output);
             Util.WriteLine(output);
         }
 
-        public static void InfoMate(int depth, int seldepth, int mateIn, long nodes, long timeMs, ulong[] pv, int hashfull, long tbHits)
+        public void InfoMate(int depth, int seldepth, int mateIn, long nodes, long timeMs,
+            ulong[] pv, int hashfull, long tbHits)
         {
+            if (!enable)
+            {
+                return;
+            }
             StringBuilder sb = new();
             int nps = (int)(nodes * 1000 / Math.Max(1, timeMs));
-            sb.Append(@$"info depth {depth} seldepth {seldepth} score mate {mateIn} nodes {nodes} hashfull {hashfull} nps {nps} time {timeMs} ");
+            sb.Append($"info depth {depth} seldepth {seldepth} score mate {mateIn} nodes {nodes} hashfull {hashfull} nps {nps} time {timeMs}");
             if (tbHits > 0)
             {
-                sb.Append($@"tbhits {tbHits} ");
+                sb.Append($" tbhits {tbHits}");
             }
-            sb.Append("pv");
+            sb.Append(" pv");
             for (int n = 0; n < pv.Length; n++)
             {
-                sb.Append($@" {Move.ToString(pv[n])}");
+                sb.Append($" {Move.ToString(pv[n])}");
             }
-
             string output = sb.ToString();
-            if (!DisableOutput)
-            {
-                Console.Out.WriteLineAsync(output);
-            }
+            Console.Out.WriteLineAsync(output);
             Util.WriteLine(output);
         }
 
-        public static void BestMove(ulong bestmove, ulong? suggestedPonder = null)
+        public void BestMove(string bestmove, string? suggestedPonder = null)
         {
-            Console.Write(@$"bestmove {Move.ToString(bestmove)}");
-            Util.Write(@$"bestmove {Move.ToString(bestmove)}");
-            if (suggestedPonder.HasValue)
+            if (!enable)
             {
-                Console.WriteLine(@$" ponder {Move.ToString(suggestedPonder.Value)}");
-                Util.WriteLine(@$" ponder {Move.ToString(suggestedPonder.Value)}");
+                return;
             }
-            else
-            {
-                Console.WriteLine();
-                Util.WriteLine();
-            }
-        }
-
-        public static void BestMove(string bestmove, string? suggestedPonder = null)
-        {
-            Console.Write(@$"bestmove {bestmove}");
-            Util.Write(@$"bestmove {bestmove}");
-
+            StringBuilder sb = new();
+            sb.Append($"bestmove {bestmove}");
             if (suggestedPonder != null)
             {
-                Console.Out.WriteLineAsync(@$" ponder {suggestedPonder}");
-                Util.WriteLine(@$" ponder {suggestedPonder}");
+                sb.Append($" ponder {suggestedPonder}");
             }
-            else
+            string output = sb.ToString();
+            Console.Out.WriteLine(output);
+            Util.WriteLine(output);
+        }
+
+        public void BestMove(ulong bestmove, ulong? suggestedPonder = null)
+        {
+            if (!enable)
             {
-                Console.Out.WriteLineAsync();
-                Util.WriteLine();
+                return;
             }
+
+            string? ponder = suggestedPonder.HasValue ? Move.ToString(suggestedPonder.Value) : null;
+            BestMove(Move.ToString(bestmove), ponder);
         }
 
-        public static void CurrentMove(int depth, ulong move, int moveNumber, long nodes, int hashfull)
+        public void CurrentMove(int depth, ulong move, int moveNumber, long nodes, int hashfull)
         {
-            Console.Out.WriteLineAsync($@"info depth {depth} currmove {Move.ToString(move)} currmovenumber {moveNumber} nodes {nodes} hashfull {hashfull}");
+            if (!enable)
+            {
+                return;
+            }
+            string output = $"info depth {depth} currmove {Move.ToString(move)} currmovenumber {moveNumber} nodes {nodes} hashfull {hashfull}";
+            Console.Out.WriteLineAsync(output);
         }
 
-        public static void Usage(int cpuload)
+        public void Usage(int cpuload)
         {
-            Console.Out.WriteLineAsync($@"info cpuload {cpuload}");
+            if (!enable)
+            {
+                return;
+            }
+            Console.Out.WriteLineAsync($"info cpuload {cpuload}");
         }
+
+        public static Uci Default => defaultUci;
+
+        private bool enable;
+        private bool debug;
+        private static Uci defaultUci;
     }
 }
