+++ Pedantic.Chess/UciOptions.cs
@@ -17,6 +17,7 @@ namespace Pedantic.Chess
         public const bool DEFAULT_SYZYGY_PROBE_ROOT = true;
         public const int DEFAULT_SYZYGY_PROBE_DEPTH = 2;
         public const bool DEFAULT_ANALYSE_MODE = false;
+        public const int DEFAULT_THREADS = 1;
 
         static UciOptions()
         {
@@ -29,6 +30,7 @@ namespace Pedantic.Chess
             SyzygyProbeRoot = DEFAULT_SYZYGY_PROBE_ROOT;
             SyzygyProbeDepth = DEFAULT_SYZYGY_PROBE_DEPTH;
             AnalyseMode = DEFAULT_ANALYSE_MODE;
+            Threads = DEFAULT_THREADS;
         }
 
         public static bool CollectStatistics { get; set; }
@@ -40,5 +42,6 @@ namespace Pedantic.Chess
         public static bool SyzygyProbeRoot { get; set; }
         public static int SyzygyProbeDepth { get; set; }
         public static bool AnalyseMode { get; set; }
+        public static int Threads { get; set; }
     }
 }
