--- Pedantic.Genetics/ChessWeights.cs
@@ -1,1405 +1,0 @@
-﻿// ***********************************************************************
-// Assembly         : Pedantic.Genetics
-// Author           : JoAnn D. Peeler
-// Created          : 03-12-2023
-//
-// Last Modified By : JoAnn D. Peeler
-// Last Modified On : 03-28-2023
-// ***********************************************************************
-// <copyright file="ChessWeights.cs" company="Pedantic.Genetics">
-//     Copyright (c) . All rights reserved.
-// </copyright>
-// <summary>
-//     Mapped POCO class for maintaining the weights in the LiteDB 
-//     database.
-// </summary>
-// ***********************************************************************
-using Pedantic.Utilities;
-
-namespace Pedantic.Genetics
-{
-    public sealed class ChessWeights
-    {
-        public static readonly Guid DEFAULT_IMMORTAL_ID = new("da5e310e-b0dc-4c77-902c-5a46cc81bb73");
-        public const int MAX_WEIGHTS = 3978;
-        public const int ENDGAME_WEIGHTS = 1989;
-        public const int PIECE_WEIGHT_LENGTH = 6;
-        public const int PIECE_SQUARE_LENGTH = 1536;
-        public const int PIECE_VALUES = 0;
-        public const int PIECE_SQUARE_TABLE = 6;
-        public const int PIECE_MOBILITY = 1542;
-        public const int KING_ATTACK = 1546;
-        public const int PAWN_SHIELD = 1549;
-        public const int ISOLATED_PAWN = 1552;
-        public const int UNUSED_0 = 1553;
-        public const int DOUBLED_PAWN = 1554;
-        public const int CONNECTED_PAWN = 1555;
-        public const int KNIGHT_OUTPOST = 1619;
-        public const int BISHOP_OUTPOST = 1620;
-        public const int BISHOP_PAIR = 1621;
-        public const int ROOK_ON_OPEN_FILE = 1622;
-        public const int ROOK_ON_HALF_OPEN_FILE = 1623;
-        public const int ROOK_BEHIND_PASSED_PAWN = 1624;
-        public const int DOUBLED_ROOKS_ON_FILE = 1625;
-        public const int KING_ON_OPEN_FILE = 1626;
-        public const int KING_ON_HALF_OPEN_FILE = 1627;
-        public const int CASTLING_AVAILABLE = 1628;
-        public const int CASTLING_COMPLETE = 1629;
-        public const int CENTER_CONTROL = 1630;
-        public const int QUEEN_ON_OPEN_FILE = 1632;
-        public const int QUEEN_ON_HALF_OPEN_FILE = 1633;
-        public const int ROOK_ON_7TH_RANK = 1634;
-        public const int PASSED_PAWN = 1635;
-        public const int BAD_BISHOP_PAWN = 1699;
-        public const int BLOCK_PASSED_PAWN = 1763;
-        public const int SUPPORTED_PAWN = 1811;
-        public const int KING_OUTSIDE_PP_SQUARE = 1875;
-        public const int PP_FRIENDLY_KING_DISTANCE = 1876;
-        public const int PP_ENEMY_KING_DISTANCE = 1877;
-        public const int PAWN_RAM = 1878;
-        public const int PIECE_THREAT = 1942;
-        public const int PAWN_PUSH_THREAT = 1978;
-        public const int KING_ON_OPEN_DIAGONAL = 1984;
-        public const int PP_CAN_ADVANCE = 1985;
-
-        public ChessWeights(Guid _id, bool isActive, bool isImmortal, string description, short[] weights, 
-            float fitness, int sampleSize, float k, short totalPasses, DateTime updatedOn, DateTime createdOn)
-        {
-            Id = _id;
-            IsActive = isActive;
-            IsImmortal = isImmortal;
-            Description = description;
-            Weights = weights;
-            Fitness = fitness;
-            SampleSize = sampleSize;
-            K = k;
-            TotalPasses = totalPasses;
-            UpdatedOn = updatedOn;
-            CreatedOn = createdOn;
-        }
-
-        public ChessWeights(ChessWeights other)
-        {
-            Id = Guid.NewGuid();
-            IsActive = other.IsActive;
-            IsImmortal = other.IsImmortal;
-            Description = other.Description;
-            Weights = ArrayEx.Clone(other.Weights);
-            Fitness = other.Fitness;
-            SampleSize = other.SampleSize;
-            K = other.K;
-            TotalPasses = other.TotalPasses;
-            UpdatedOn = DateTime.UtcNow;
-            CreatedOn = DateTime.UtcNow;
-        }
-
-        public ChessWeights(short[] weights)
-        {
-            Id = Guid.NewGuid();
-            IsActive = true;
-            IsImmortal = false;
-            Description = "Anonymous";
-            Weights = ArrayEx.Clone(weights);
-            Fitness = 0;
-            SampleSize = 0;
-            K = 0;
-            TotalPasses = 0;
-            UpdatedOn = DateTime.UtcNow;
-            CreatedOn = DateTime.UtcNow;
-        }
-
-        public ChessWeights()
-        {
-            Id = Guid.NewGuid();
-            IsActive = false;
-            IsImmortal = false;
-            Description = string.Empty;
-            Weights = Array.Empty<short>();
-            Fitness = default;
-            SampleSize = default;
-            K = default;
-            TotalPasses = default;
-            UpdatedOn = DateTime.UtcNow;
-            CreatedOn = DateTime.UtcNow;
-        }
-
-        public Guid Id { get; set; }
-        public bool IsActive { get; set; }
-        public bool IsImmortal { get; set; }
-        public string Description { get; set; }
-        public short[] Weights { get; init; }
-        public float Fitness { get; set; }
-        public int SampleSize { get; set; }
-        public float K { get; set; }
-        public short TotalPasses { get; set; }
-        public DateTime UpdatedOn { get; set; }
-
-        public DateTime CreatedOn { get; set; }
-
-        public static ChessWeights Empty { get; } = new ChessWeights();
-
-        public static ChessWeights CreateParagon()
-        {
-            ChessWeights paragon = new(paragonWeights)
-            {
-                Id = DEFAULT_IMMORTAL_ID,
-                IsImmortal = true,
-                Fitness = 0,
-                SampleSize = 0,
-                K = 0,
-                TotalPasses = 0,
-                UpdatedOn = DateTime.UtcNow
-            };
-            return paragon;
-        }
-
-        public static bool LoadParagon(out ChessWeights paragon)
-        {
-            var rep = new ChessDb();
-            ChessWeights? p = rep.Weights.GetAll().FirstOrDefault(w => w.IsActive && w.IsImmortal);
-            if (p == null)
-            {
-                p = CreateParagon();
-                rep.Weights.Insert(p);
-                rep.Save();
-                paragon = p;
-            }
-            else
-            {
-                paragon = p;
-            }
-            return true;
-        }
-
-#if false
-        public static Score S(short mg, short eg) => new Score(mg, eg);
-
-        // Solution sample size: 3970, generated on Sun, 24 Sep 2023 11:46:12 GMT
-        private static readonly Score[] defaultWeights =
-        {
-            /* piece values */
-            S(116, 155),   S(508, 452),   S(556, 500),   S(695, 867),   S(1587, 1528), S(  0,   0),
-
-            /* piece square values */
-            # region piece square values
-
-            /* pawns: KK */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-64,  13),   S(-59,  11),   S(-44,  -4),   S(-27, -25),   S(-53,  28),   S( 52, -11),   S( 60, -34),   S(-13, -39),
-            S(-67,   6),   S(-56,  -7),   S(-58, -26),   S(-70, -17),   S(-44, -26),   S(-16, -24),   S(  7, -28),   S(-55, -19),
-            S(-33,   9),   S(-36,   4),   S(-11, -37),   S(-33, -50),   S( -8, -38),   S( 25, -39),   S(  3, -16),   S( -4, -30),
-            S(-15,  57),   S(-31,  29),   S(  0,   9),   S( 21,  -9),   S( 36, -34),   S( 23, -23),   S( 43,  -6),   S( 59, -15),
-            S( 51,  85),   S( -4,  97),   S( -6,  88),   S( 69,  72),   S(121,  67),   S(194,  18),   S(187,  32),   S(179,  21),
-            S( 92,  95),   S(105,  92),   S( 81, 101),   S(117,  55),   S( 95,  64),   S(157, -11),   S( 70,  54),   S( 96,   3),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* pawns: KQ */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-100,  29),  S(-99,  37),   S(-70,  15),   S(-78,  48),   S(-49, -21),   S( 47, -40),   S( 62, -46),   S(-16, -19),
-            S(-90,   3),   S(-91,   4),   S(-73, -24),   S(-81, -29),   S(-44, -49),   S(-22, -37),   S( 13, -41),   S(-52, -11),
-            S(-65,  21),   S(-41,   4),   S(-27, -20),   S(-36, -40),   S(-18, -36),   S(  2, -34),   S( -1, -26),   S(  7, -28),
-            S(-52,  70),   S(-13,   2),   S(  3,  -9),   S( 15,   4),   S(  4, -15),   S(-26,  25),   S( 22,  13),   S( 58,   4),
-            S(102,  52),   S( 93,  11),   S(137, -15),   S(107,  46),   S(117, 104),   S( 64, 125),   S( 66, 126),   S(115, 109),
-            S(164,   8),   S( 89,  38),   S(131,  -3),   S( 95,  77),   S(111, 107),   S( 65, 153),   S(144, 119),   S(148, 107),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* pawns: QK */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-38,  14),   S(  4,  -3),   S( -7, -10),   S(-45,  15),   S(-66,  13),   S( -9,  -7),   S(-29,  -4),   S(-75,  18),
-            S(-29,  10),   S( -2, -15),   S(-49, -19),   S(-60, -31),   S(-57, -24),   S(-32, -31),   S(-50, -23),   S(-110,   5),
-            S(  1,   7),   S(  2,   4),   S(-12, -17),   S(-32, -36),   S(-29, -23),   S(  3, -35),   S(-21, -24),   S(-46, -13),
-            S( 17,  62),   S(-18,  55),   S(  3,  34),   S( 28,  11),   S( 18, -20),   S(  4, -25),   S( 10, -14),   S(  1,   9),
-            S( 70, 105),   S( -1, 142),   S(  5, 137),   S( 74,  98),   S(127,  62),   S(175,   7),   S(136,  33),   S( 93,  57),
-            S(123,  83),   S(190, 102),   S(120, 151),   S(160,  93),   S(100,  57),   S(122,  -4),   S( 71,  21),   S( 49,  27),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* pawns: QQ */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-56,   0),   S( -6,  -9),   S( 33, -29),   S(  2, -19),   S(-14, -23),   S(  7, -39),   S( -6, -37),   S(-29, -24),
-            S(-49,  -9),   S(-25, -13),   S(-25, -35),   S(-50, -44),   S(-21, -52),   S(-40, -54),   S(-39, -46),   S(-81, -21),
-            S(-24,   3),   S(-23,  -2),   S( 18, -41),   S( -8, -48),   S(  8, -52),   S( 12, -63),   S( -7, -41),   S(-32, -33),
-            S( 57,  12),   S( 59, -12),   S( 60, -26),   S( 54, -10),   S( 28, -35),   S( -8, -23),   S( 19, -27),   S( 35, -14),
-            S(226, -15),   S(145,  -3),   S(176, -31),   S(148,  23),   S( 45,  81),   S( 63,  57),   S( 19,  68),   S(115,  56),
-            S(177,  -7),   S( 71,  50),   S(109, -13),   S( 64,  62),   S(110,  36),   S( 22,  80),   S(136,  55),   S(109,  65),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* knights: KK */
-            S(-113, -41),  S(-21, -75),   S(-48,   0),   S(-28,  16),   S( -8,   8),   S( -5,  10),   S(-19, -10),   S(-64, -57),
-            S(-34, -41),   S(-47,  27),   S( -5,   7),   S( 10,  12),   S( 11,  21),   S( 21,  33),   S( -3,  25),   S( -4,   7),
-            S(-24,  -9),   S( -6,  23),   S(  5,  33),   S( 23,  61),   S( 33,  54),   S( 13,  28),   S(  9,  39),   S(  3,  34),
-            S( -6,  33),   S( 15,  54),   S( 26,  79),   S( 25,  91),   S( 31,  99),   S( 50,  81),   S( 54,  72),   S( 17,  45),
-            S( 16,  48),   S( 20,  48),   S( 46,  61),   S( 50,  91),   S( 19, 101),   S( 46, 107),   S( 17, 113),   S( 69,  54),
-            S(-31,  67),   S( 26,  49),   S( 45,  59),   S( 54,  60),   S( 96,  45),   S(181,  19),   S( 40,  75),   S( -9,  66),
-            S(-13,  40),   S(-24,  62),   S(  7,  63),   S( 62,  61),   S( 48,  67),   S(134,   1),   S( 17,  24),   S(-14,  45),
-            S(-309,  58),  S(-53,  69),   S(-30,  69),   S( 72,  47),   S( 73,  44),   S(-110,  50),  S(  4,  47),   S(-207, -75),
-
-            /* knights: KQ */
-            S(-96, -110),  S(-46, -48),   S(-63,   2),   S(-54,  13),   S(-34, -12),   S(-27, -15),   S(-43,  15),   S(-25, -127),
-            S(-85, -15),   S(-41,  11),   S(-12, -21),   S( 14, -17),   S( -4,  -5),   S( 20,   1),   S(-68,  32),   S(-51, -10),
-            S(-47,  -7),   S( -3, -30),   S( -4,   5),   S( 27,   3),   S( 23,   2),   S( 26, -33),   S( -9,  14),   S(-20,  12),
-            S( -6,   3),   S( 51, -12),   S( 38,  19),   S( 62,   8),   S( 51,  18),   S( 29,  26),   S( 27,  15),   S(-16,  22),
-            S( 63, -46),   S( 30, -21),   S( 71,  -4),   S( 69,   5),   S( 89,  -2),   S( 70,   2),   S( 53,  -5),   S( 21,  -6),
-            S( 55, -39),   S( 76, -65),   S(146, -70),   S(136, -51),   S( 74,  -5),   S( 74,   0),   S( -4,  30),   S( -4,   1),
-            S(  7, -45),   S(-14, -29),   S( 10, -44),   S( 80, -17),   S( 33,  -7),   S( 33,   7),   S(-26,  22),   S(-42,   1),
-            S(-294, -119), S(-13, -55),   S(-11, -35),   S( 39,  -8),   S(-43, -30),   S( -2,  36),   S(-32,  34),   S(-122, -54),
-
-            /* knights: QK */
-            S(-101, -104), S(-76,  23),   S(-52,  21),   S(-74,  45),   S(-31,  -4),   S(-25, -19),   S(-57, -36),   S(-94, -80),
-            S(-44, -61),   S(-32,  16),   S( 12,  -4),   S( -6,  10),   S(-13,   8),   S( -1,   6),   S(-63,   5),   S(-17, -57),
-            S(-16,  10),   S( -3, -13),   S( 25, -11),   S( 48,  21),   S( 61,  -2),   S(-18,  16),   S(  2, -14),   S(-56, -27),
-            S( 28,   0),   S( 56,  11),   S( 53,  22),   S( 38,  49),   S( 51,  22),   S( 39,  28),   S( 39,   5),   S(-13,   1),
-            S( 19,  25),   S( 61,  12),   S( 77,  22),   S(126,  -1),   S( 74,  24),   S( 85,  14),   S( 56,   3),   S( 74, -23),
-            S( -1,  35),   S( 15,  31),   S( 45,  30),   S( 87,  -9),   S(109,  -6),   S(166, -39),   S( 46,  -9),   S( 24, -26),
-            S(-24,  26),   S(  4,  13),   S( 21,  38),   S( 73,  42),   S(-12,  34),   S( 74, -10),   S( -3, -20),   S( 65, -36),
-            S(-148, -84),  S(-34,  22),   S(-23,  36),   S( 29,  -5),   S( -7,  25),   S(-35, -26),   S(-34, -27),   S(-59, -99),
-
-            /* knights: QQ */
-            S(  8, -108),  S(-37, -15),   S( 30, -30),   S(-73,   9),   S(-40,  -2),   S(-48,   3),   S(-47, -72),   S(-67, -83),
-            S(-36, -41),   S(-34,  28),   S( 14, -31),   S( 29, -20),   S(-17, -10),   S( 14, -53),   S(  1, -19),   S(-139, -39),
-            S( 10, -23),   S(  9, -24),   S( 24, -27),   S( 38,  -6),   S( 31,  -6),   S(-12, -17),   S(-14,  -4),   S(-30, -48),
-            S( 17,  43),   S( 25,   7),   S( 79,   0),   S( 82, -15),   S( 52,  -6),   S( 50, -15),   S( 52,  -1),   S(-35,  11),
-            S( 50,  -9),   S( 24,   9),   S( 71, -17),   S( 87, -15),   S( 86,  -8),   S( 72, -21),   S( 74, -29),   S( 13,   2),
-            S( 78, -24),   S(-35,  -1),   S( 52, -28),   S(148, -38),   S( 64, -23),   S( 39,   9),   S( 71, -26),   S(  5, -26),
-            S(-21,  30),   S( 24, -12),   S( -7,  -2),   S( 21,  25),   S( 54,  10),   S( 52,   1),   S(-12,  10),   S(-57,  16),
-            S(-121, -86),  S(  0,  58),   S( -1,   6),   S(  1,  30),   S(-20, -15),   S( 16,  33),   S(  5,  48),   S(-50, -45),
-
-            /* bishops: KK */
-            S( 19, -15),   S( 20, -16),   S(  2,   3),   S(-19,  46),   S(  6,  37),   S(  3,  42),   S(  6,  11),   S(  5, -49),
-            S( 21,   2),   S( 21,  11),   S( 19,  22),   S( 12,  44),   S( 16,  39),   S( 27,  32),   S( 45,   7),   S( 30, -45),
-            S( -1,  26),   S( 20,  42),   S( 31,  62),   S( 21,  52),   S( 19,  76),   S( 21,  47),   S( 23,  22),   S(  6,  21),
-            S( -6,  48),   S(  1,  63),   S( 21,  72),   S( 49,  57),   S( 39,  46),   S( 19,  51),   S( 10,  36),   S( 35,  -9),
-            S( -2,  65),   S( 24,  36),   S( 15,  50),   S( 56,  48),   S( 37,  56),   S( 51,  49),   S( 26,  45),   S(  7,  59),
-            S(  4,  50),   S( 20,  58),   S( -4,  58),   S(  1,  57),   S( 13,  65),   S( 84,  62),   S( 58,  48),   S( -7,  88),
-            S(-44,  75),   S(-22,  65),   S( -7,  66),   S(-24,  76),   S(-29,  74),   S(  8,  54),   S(-79,  62),   S(-57,  50),
-            S(-100,  95),  S(-61,  86),   S(-36,  83),   S(-36,  89),   S(-48,  83),   S(-83,  71),   S( 10,  28),   S(-70,  56),
-
-            /* bishops: KQ */
-            S(-50,   2),   S( 28,  -9),   S(-28,  13),   S(-19,   5),   S( 18,  -9),   S(  6, -14),   S(102, -91),   S( 16, -95),
-            S(-21, -15),   S( -7,  -1),   S( -2,  12),   S( 10,   8),   S( 19,   3),   S( 51, -24),   S( 84, -27),   S( 19, -70),
-            S(-29,   8),   S(  7,   8),   S( 31,  13),   S( 32,   5),   S( 29,  15),   S( 58, -10),   S( 59, -16),   S( 48, -37),
-            S( 22,  15),   S(  4,  22),   S( 18,   4),   S( 68, -10),   S( 85,  -6),   S( 45,  24),   S( 23,  17),   S( -6,  19),
-            S( 11,  22),   S( 55, -12),   S( 44, -13),   S(135, -61),   S( 91,  -1),   S( 35,  22),   S( 37,  18),   S( 27,   5),
-            S(  8, -10),   S(110, -37),   S(134, -29),   S( 14,  14),   S( 47,   5),   S( 42,  12),   S( 31,  23),   S(  6,  35),
-            S(-49, -32),   S( 32, -17),   S(-43,  22),   S(-29,  18),   S(  3,  11),   S( -6,  21),   S(-18,  44),   S(-61,  69),
-            S(-32,   5),   S(-34,   6),   S( 18,  -6),   S(-15,  21),   S(-12,  19),   S(-21,  35),   S(-27,  38),   S(-50,  36),
-
-            /* bishops: QK */
-            S( 40, -65),   S( 34, -35),   S( 21,   4),   S(-25,  27),   S( -8,  18),   S(-28,   1),   S(-30,  27),   S(-134,   0),
-            S( 60, -57),   S( 70, -22),   S( 30,  12),   S( 36,  15),   S( -4,  15),   S(-14,  14),   S(-29,  -2),   S(-11, -19),
-            S( 14,  -2),   S( 47,   8),   S( 51,  16),   S( 28,  27),   S( 19,  35),   S(  6,   4),   S( 10,  -7),   S(-17, -15),
-            S(  4,   2),   S( 33,  21),   S( 32,  41),   S( 83,   4),   S( 60,  -5),   S( 13,  28),   S(-25,  30),   S(-11,  19),
-            S(-10,  47),   S( 30,  22),   S( 47,  17),   S( 98,  -3),   S( 54,   7),   S( 56,  13),   S( 50, -13),   S(-14,  31),
-            S(-30,  47),   S(  0,  47),   S( 46,  16),   S( 34,  32),   S( 57,  25),   S( 79,  20),   S( 91, -29),   S( 21,   8),
-            S(-40,  53),   S(-57,  45),   S(-34,  51),   S(-24,  50),   S(-60,  50),   S(  3,  29),   S(-26,  -3),   S(-29, -27),
-            S(-56,  35),   S(-56,  40),   S(-24,  42),   S(-44,  55),   S(  4,  14),   S(-33,   2),   S( 26,   7),   S(-28,  18),
-
-            /* bishops: QQ */
-            S(-78, -19),   S(-24, -17),   S( 19, -32),   S(-61,  -9),   S(  0, -20),   S(-15, -34),   S( 38, -51),   S(-19, -82),
-            S( 49, -74),   S( 68, -49),   S( -9,  -9),   S( 41, -24),   S(-28,   4),   S( 38, -45),   S( -6, -33),   S( 92, -107),
-            S( 12, -21),   S( 42, -27),   S( 48, -24),   S( 22, -20),   S( 39, -17),   S( 22, -27),   S( 51, -57),   S( -4, -24),
-            S( 49, -19),   S( 60, -21),   S( 24, -21),   S( 87, -43),   S( 34, -29),   S( 70, -26),   S( 19, -23),   S( 25, -32),
-            S( 34, -16),   S( 38, -21),   S( 37, -33),   S( 95, -48),   S( 96, -56),   S( 11, -21),   S( 37, -36),   S(  6,  -7),
-            S(-36,  -3),   S( 92, -24),   S( 98, -37),   S( 62, -25),   S(  4,   3),   S( 71, -38),   S( -4,  -7),   S(  1,  -6),
-            S(-47, -23),   S(-24, -14),   S( 36, -24),   S(  0,   5),   S( 30, -23),   S(-19, -11),   S( 36, -18),   S(-19,   4),
-            S(-27, -15),   S( -5, -11),   S( -4, -20),   S( -1,   8),   S( 20,   9),   S( -8,  -4),   S(  7,   9),   S(-30,  35),
-
-            /* rooks: KK */
-            S(-33, 100),   S(-21,  93),   S(-16,  80),   S( -3,  53),   S(  5,  38),   S( -1,  65),   S( 13,  60),   S(-28,  51),
-            S(-49,  97),   S(-32,  89),   S(-25,  77),   S(-15,  52),   S(-13,  39),   S( 11,  24),   S( 37,  38),   S(-18,  52),
-            S(-50, 100),   S(-47, 103),   S(-34,  78),   S(-21,  52),   S(  1,  30),   S( -4,  44),   S( 30,  34),   S( -2,  43),
-            S(-43, 123),   S(-31, 109),   S(-22,  93),   S(-10,  72),   S(-19,  55),   S(-26,  86),   S( 28,  71),   S(-21,  74),
-            S(-24, 134),   S( -8, 118),   S(  2,  98),   S( 26,  67),   S( -1,  85),   S( 40,  71),   S( 73,  74),   S( 24,  82),
-            S(-10, 133),   S( 28, 111),   S( 35,  94),   S( 38,  67),   S( 73,  49),   S(140,  52),   S(132,  65),   S( 68,  74),
-            S(-30, 128),   S(-29, 130),   S( -2,  94),   S( 33,  62),   S( 19,  69),   S( 87,  65),   S(102,  88),   S(141,  62),
-            S( 57, 104),   S( 45, 120),   S( 39,  93),   S( 56,  61),   S( 36,  94),   S( 98,  96),   S(115, 108),   S(126,  81),
-
-            /* rooks: KQ */
-            S(-51,  15),   S( -4, -17),   S(-16, -26),   S(-25, -11),   S(-24, -10),   S(-18,  -8),   S(-31,  34),   S(-59,  42),
-            S(-70,  12),   S(-17, -32),   S(-31, -30),   S(-29, -27),   S(-26, -19),   S(-37,   3),   S(-42,  21),   S(-42,  11),
-            S(-49,   2),   S(-26, -16),   S(-45, -19),   S(-18, -39),   S(-52, -13),   S(-27,  -9),   S( -6,   1),   S(-50,   7),
-            S(-68,  15),   S(-14, -10),   S(-65,   4),   S( -7, -20),   S(-20, -18),   S(-50,  27),   S( -7,  17),   S( -7,   6),
-            S(-10,   7),   S( 46, -16),   S( -8,  -4),   S( -8,   5),   S( 43, -26),   S( 32,  -2),   S( 41,  12),   S(  7,  24),
-            S( 79, -11),   S( 90, -20),   S( 56, -34),   S( 79, -41),   S( 77, -42),   S( 55,  -6),   S(131, -20),   S( 53,  14),
-            S(-19,  20),   S( 17,   8),   S( 53, -32),   S( 47, -25),   S( 73, -28),   S( 13,  -5),   S( 50,  14),   S( 96,  -8),
-            S( 19,  39),   S( 50,  14),   S( 53, -13),   S( 45, -14),   S( 40,  -9),   S( 75,  -2),   S( 74,  30),   S(110,   0),
-
-            /* rooks: QK */
-            S(-93,  54),   S(-27,  26),   S(-32,  22),   S(-24, -11),   S(-17, -25),   S(-16, -25),   S( -4, -26),   S(-45,  12),
-            S(-48,  20),   S(-48,  17),   S(-50,  23),   S(-36, -12),   S(-51, -19),   S(-51, -23),   S(-23, -28),   S(-35, -15),
-            S(-39,  17),   S( -5,  16),   S(-33,   8),   S(-44,  -9),   S(-67,  -1),   S(-73,  -7),   S(-23, -32),   S(-32, -30),
-            S( -5,  10),   S( 43,  -2),   S(-33,  22),   S(-34,   8),   S(-52,   9),   S(-51,  11),   S(  9,  -5),   S(-24,  -3),
-            S( 62,  11),   S( 33,  33),   S( 60,   0),   S(-15,  20),   S(-24,  12),   S(-10,   7),   S( 17,   4),   S(  6, -12),
-            S( 44,  26),   S(100,  11),   S( 40,  24),   S(-13,  18),   S( 55, -17),   S( 99, -11),   S(106, -13),   S( 46,   2),
-            S( 35,  18),   S( 45,  20),   S( 78,  -2),   S(  3,  10),   S(-35,  20),   S( 77, -10),   S( 33,  13),   S( 12,  16),
-            S(242, -106),  S(103,  -1),   S( 80,   3),   S( 39,  14),   S( 11,  19),   S( 72,  -3),   S( 93,   6),   S( 94,  -3),
-
-            /* rooks: QQ */
-            S(-56, -15),   S(-21, -33),   S( 13, -69),   S( 17, -87),   S( -4, -63),   S(-18, -30),   S(-34,  -5),   S(-44,  15),
-            S(-32, -47),   S(-40, -26),   S(-33, -45),   S(  8, -86),   S(-29, -74),   S(-44, -56),   S(-26, -44),   S(-29, -17),
-            S(-80, -14),   S( 27, -54),   S(  3, -65),   S(-12, -72),   S(-37, -36),   S(-51, -49),   S(-47, -12),   S(-52,  -4),
-            S(-39, -13),   S( 45, -46),   S(-32, -26),   S(-15, -45),   S(-47, -48),   S(-26, -33),   S(-49,  -8),   S(-31,   4),
-            S( 39, -31),   S( 69, -42),   S(  7, -30),   S(108, -79),   S( 40, -50),   S( 38, -34),   S(  2,   0),   S(  6,   1),
-            S( 79, -38),   S( 92, -29),   S(117, -49),   S( 93, -63),   S( 69, -63),   S( 67, -30),   S( 31, -10),   S( 18,   1),
-            S( 89, -40),   S( 65, -24),   S( 21, -12),   S( 78, -57),   S(158, -94),   S( 63, -58),   S( 48, -22),   S( 13,   0),
-            S( 43,   2),   S( 32,  11),   S(  6,  -4),   S( -8,  -4),   S(  9, -38),   S( 55, -33),   S( 97,  -2),   S( 13,   9),
-
-            /* queens: KK */
-            S( 39,  56),   S( 46,  30),   S( 56,   4),   S( 60,  15),   S( 80, -49),   S( 60, -76),   S(-11, -25),   S(  3,  56),
-            S( 26,  84),   S( 49,  53),   S( 50,  49),   S( 59,  31),   S( 61,  21),   S( 87, -31),   S( 84, -50),   S( 50, -16),
-            S( 23,  61),   S( 38,  62),   S( 45,  70),   S( 33,  70),   S( 51,  53),   S( 38,  86),   S( 50,  80),   S( 47,  30),
-            S( 33,  83),   S( 15, 105),   S( 24,  89),   S( 20,  97),   S( 20,  87),   S( 31, 106),   S( 54,  87),   S( 16, 139),
-            S( 24,  95),   S( 32, 100),   S( -8, 138),   S( -4, 128),   S(  5, 141),   S( 17, 170),   S( 24, 184),   S( 21, 197),
-            S( 15, 111),   S( 22, 147),   S(-14, 151),   S(-31, 162),   S( 10, 165),   S(105, 156),   S( 56, 205),   S( 18, 251),
-            S(-15, 165),   S(-33, 196),   S(-27, 183),   S(-15, 156),   S(-56, 206),   S( 80, 141),   S( 19, 280),   S(126, 167),
-            S(-34, 166),   S( 33, 128),   S( 16, 135),   S(  3, 155),   S( 27, 152),   S(108, 144),   S(159, 154),   S(125, 173),
-
-            /* queens: KQ */
-            S( 18, -26),   S( 14, -13),   S( 28, -90),   S( 19, -47),   S( 20, -26),   S(  7, -60),   S(-29, -26),   S(-25,  15),
-            S( 33,   8),   S( 31, -34),   S( 45, -85),   S( 23, -35),   S( 29, -37),   S( 21, -60),   S(-37, -28),   S(-51, -48),
-            S(  9,   1),   S(  7, -20),   S( 41, -86),   S( 23, -34),   S( 32, -30),   S( 27,  19),   S( 13, -12),   S( -5, -57),
-            S(  1,  32),   S( 57, -78),   S(-20,  10),   S(-14,  -9),   S( 16, -11),   S( -9,  29),   S(  3,  29),   S(-14,  51),
-            S( 19, -11),   S( 32, -14),   S(  1, -10),   S( -3, -18),   S( 40, -35),   S( -5,  28),   S(  6,  39),   S(  5,  69),
-            S( 52,  -2),   S( 92,   4),   S( 62,  -1),   S( 25,  -5),   S( 11,  18),   S( 19,  38),   S( 28,  70),   S( 16,  47),
-            S( 57,  26),   S( 95,  10),   S( 71,  17),   S( 73,   4),   S( 39,  -2),   S(-11,  71),   S(-17,  90),   S( 22,  51),
-            S( 78,  39),   S(124,  35),   S( 92,  34),   S( 80,  52),   S( 25,  67),   S( 56,  28),   S( 49,  24),   S(-15,  24),
-
-            /* queens: QK */
-            S(-74, -56),   S(-87, -31),   S( -7, -106),  S( -5, -70),   S( 16, -103),  S( -2, -61),   S(-28, -28),   S(-59,  31),
-            S(-68, -24),   S(-35, -85),   S( 39, -116),  S( 23, -72),   S( 21, -72),   S(  1, -52),   S( 41, -64),   S(  2,   0),
-            S(-47, -33),   S( -1, -22),   S(  6, -10),   S(-12,  11),   S( 14, -35),   S(-18,  -5),   S( 14, -50),   S( -3,  -1),
-            S(-16, -31),   S(-42,  33),   S(-28,  42),   S( -7,   3),   S(-19,  12),   S( -3, -17),   S( 11, -20),   S(-13,  -5),
-            S( -8,  27),   S(-13,  11),   S(-44,  67),   S(-38,  45),   S(-32,  27),   S( 39, -18),   S(  4,  20),   S( 13, -23),
-            S( -2,  17),   S( -4,  73),   S(-19,  45),   S(-31,  47),   S( 11,  23),   S( 68,   9),   S( 40,  30),   S( 18,  25),
-            S(  6,  -7),   S( -1,  33),   S(-53,  86),   S(-44,  43),   S(-26,  27),   S( -4,  27),   S( 48,  12),   S( 61,  34),
-            S(-37,  42),   S( 11,  35),   S( 24,  26),   S( 21,  23),   S(-34,  29),   S( 75,  40),   S( 92,  31),   S( 61,  44),
-
-            /* queens: QQ */
-            S(-58, -53),   S(-74, -45),   S(-39, -15),   S(-40, -58),   S( 11, -37),   S(-14, -59),   S(-68, -31),   S(-76, -13),
-            S(-78, -30),   S(-49, -53),   S(  7, -69),   S( 24, -89),   S(-36, -21),   S(-23, -72),   S(-17, -48),   S(-62, -35),
-            S(-24, -18),   S(-14,  -8),   S( 10, -59),   S(-11, -54),   S(-12, -61),   S(-31, -35),   S(-18, -50),   S(-38, -37),
-            S(-46, -25),   S( 25, -23),   S(-39,  -3),   S(  1, -54),   S( -6, -63),   S( -2, -103),  S(-15, -61),   S(-38,   2),
-            S(-26,   3),   S(-19,  13),   S( 44, -42),   S( 37, -48),   S( 66, -50),   S(  3, -44),   S(-17, -12),   S( -1, -44),
-            S( 37,  13),   S( 46,  14),   S( 90,  39),   S( 47, -26),   S(  2, -24),   S(-33, -13),   S(-12,  -2),   S( -8,  49),
-            S( 82,  22),   S( 50,  42),   S( 67,  12),   S( 26,  32),   S(  7,  -5),   S(-22, -14),   S(-12,  20),   S(-44,   7),
-            S(  8, -35),   S( 68,  42),   S( 16,  26),   S( 24,  11),   S( 12, -10),   S( 12, -29),   S( 25, -23),   S( 13, -29),
-
-            /* kings: KK */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(-107,  -2),  S(-124,  23),  S(-49,  -7),   S(-59, -38),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(-55,  10),   S(-46,  10),   S(-25,   6),   S(-52,   6),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S( -6,  16),   S(-44,  20),   S(-49,  12),   S(-101,  12),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(112,   5),   S( 62,  17),   S( 12,  16),   S(-126,  24),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(232, -21),   S(288, -36),   S(178,  -6),   S( 34,  -2),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(253, -19),   S(297, -29),   S(269,   2),   S(153,   4),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(206, -14),   S(150,  -5),   S(147,  50),   S( 65,  13),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S( 66,   1),   S( 69,  30),   S( 67,  37),   S( -7, -140),
-
-            /* kings: KQ */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(-149,  19),  S(-164,  26),  S(-71, -14),   S(-81, -24),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(-53,   0),   S(-55,   2),   S(-60,  19),   S(-83,  12),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(-42,  15),   S(-15,   1),   S(-39,   9),   S(-115,  19),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(128,  -5),   S( 56,  20),   S(-17,  35),   S(-84,  33),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(229, -12),   S(144,  21),   S( 83,  36),   S( -3,  39),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(175,   5),   S( 96,  41),   S(113,  38),   S( -6,  57),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(182,  -3),   S(219, -10),   S(101,  27),   S( 31,  23),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S( 72,  -8),   S( 74,   2),   S( 83, -23),   S(  3, -22),
-
-            /* kings: QK */
-            S(-67, -70),   S(-45, -66),   S(-90, -43),   S(-164, -15),  S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-40, -46),   S(-60, -14),   S(-71, -29),   S(-89, -22),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-98, -16),   S(-71,  -6),   S(-39, -20),   S(-81,  -2),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-98,  13),   S(  3,   5),   S( 48,  -6),   S( 42,  -7),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-46,  12),   S( 64,  20),   S( 82,   7),   S(161, -20),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-28,  34),   S( 85,  25),   S(156,   3),   S(171, -12),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S( 45,  -4),   S(144,   4),   S(121,   2),   S(139,  -8),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-26, -21),   S( 41,  20),   S(126,  -4),   S( 53,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* kings: QQ */
-            S(-180,  47),  S(-115,   5),  S(-160,  15),  S(-204,   4),  S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-124,  13),  S(-116,  14),  S(-86,  -9),   S(-123,  -9),  S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-130,   4),  S(-74,  -7),   S(-60, -11),   S(-58,  -7),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-60, -13),   S( 16,   1),   S( 43,   2),   S(107, -13),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S( 53, -10),   S(149,  -4),   S(216, -17),   S(216, -22),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S( 65,  14),   S(171,  17),   S(165, -18),   S(124,  10),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  6,  -5),   S( 38,  63),   S( 74,  15),   S( 96,   7),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(-22, -136),  S( 26,  50),   S( 35,   9),   S( 65,   9),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            #endregion
-
-            /* mobility weights */
-            S(  9,   7), // knights
-            S(  5,   4), // bishops
-            S(  3,   3), // rooks
-            S(  1,   3), // queens
-
-            /* squares attacked near enemy king */
-            S( 24,  -5), // attacks to squares 1 from king
-            S( 21,  -2), // attacks to squares 2 from king
-            S(  7,   1), // attacks to squares 3 from king
-
-            /* pawn shield/king safety */
-            S( 20,  17), // friendly pawns 1 from king
-            S(  7,  19), // friendly pawns 2 from king
-            S(  6,  14), // friendly pawns 3 from king
-
-            /* isolated pawns */
-            S(-10,  -8),
-
-            /* backward pawns */
-            S( 15,  -6),
-
-            /* doubled pawns */
-            S(-18, -32),
-
-            /* adjacent/phalanx pawns */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S( 17,  -4),   S( -2,  -1),   S(  6,  -2),   S( 12,  37),   S( 28,  32),   S( -1, -43),   S(-30,  54),   S(  2, -47),
-            S( 13,   3),   S( 20,  -8),   S( 15,  28),   S( 19,  41),   S( 56,  -5),   S( -4,  24),   S( 18,   2),   S( 10, -18),
-            S(-17,  12),   S( 18,  22),   S( -3,  49),   S( 30,  75),   S( 35,  29),   S( 35,  38),   S( 17,   9),   S( 14,  23),
-            S( 33,  70),   S(  9,  54),   S( 66,  88),   S( 10,  86),   S(101,  73),   S(103,  37),   S( 11,  60),   S( 30,  47),
-            S( 68, 110),   S(137, 107),   S(176, 170),   S(187, 166),   S(168, 120),   S(166, 170),   S(151,  76),   S( 80,  27),
-            S( 78, 199),   S(120, 288),   S(117, 252),   S(120, 237),   S( 82, 197),   S( 53, 158),   S( 45, 212),   S( 26, 166),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* knight on outpost */
-            S( -1,  26),
-
-            /* bishop on outpost */
-            S(  6,  26),
-
-            /* bishop pair */
-            S( 40, 103),
-
-            /* rook on open file */
-            S( 35,   3),
-
-            /* rook on half-open file */
-            S( 12,  34),
-
-            /* rook behind passed pawn */
-            S(  0,  44),
-
-            /* doubled rooks on file */
-            S(  9,  15),
-
-            /* king on open file */
-            S(-79,  15),
-
-            /* king on half-open file */
-            S(-31,  28),
-
-            /* castling rights available */
-            S( 38, -42),
-
-            /* castling complete */
-            S( 19, -18),
-
-            /* center control */
-            S(  0,   9), // D0
-            S(  3,   5), // D1
-
-            /* queen on open file */
-            S(-10,  38),
-
-            /* queen on half-open file */
-            S(  7,  31),
-
-            /* rook on seventh rank */
-            S( 15,  30),
-
-            /* passed pawn */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  1,  18),   S(-15,  34),   S(-28,  33),   S(-47,  49),   S( 14,   2),   S(-30,  30),   S(-24,  60),   S( 30,  17),
-            S( 11,  25),   S( -2,  43),   S(-17,  40),   S( -9,  34),   S(-35,  46),   S(-55,  51),   S(-59,  77),   S( 19,  27),
-            S( 15,   3),   S(  5,  21),   S(-16,  32),   S( 18,  22),   S(-13,  23),   S(-40,  41),   S(-64,  68),   S(-14,  38),
-            S( 48,  30),   S( 76,  49),   S( 34,  48),   S( 15,  40),   S( 19,  69),   S( 52,  56),   S( 44,  68),   S(  4,  72),
-            S( 69,  74),   S(111,  93),   S(109,  75),   S( 34,  62),   S( -4,  56),   S( 88,  51),   S( 42,  83),   S(  2,  67),
-            S(274,  92),   S(256, 129),   S(241, 144),   S(252, 135),   S(261, 139),   S(253, 143),   S(302, 147),   S(283, 120),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* bad bishop pawns */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S( -2,  -1),   S(  1,  -9),   S( -5,  -7),   S( -4, -52),   S(-12, -17),   S(-29,   4),   S(-23,  -6),   S(  0,  -6),
-            S( -5, -10),   S( -9, -11),   S(-15, -14),   S(-10, -19),   S(-15, -19),   S(-17,  -6),   S(-16,  -8),   S( -2, -11),
-            S( -6, -12),   S(  3, -27),   S(  0, -42),   S( -1, -56),   S(-18, -35),   S(-13, -30),   S(-12, -23),   S( -2,  -7),
-            S(  5, -28),   S(  7, -38),   S(  0, -36),   S(  3, -50),   S( -7, -38),   S( -9, -31),   S( -7, -32),   S( -3, -23),
-            S( 21, -36),   S(  8, -53),   S( 17, -62),   S(  7, -62),   S( 32, -78),   S( -6, -53),   S( 49, -96),   S(-11, -19),
-            S( 65, -58),   S(107, -108),  S(103, -114),  S( 97, -136),  S( 94, -167),  S(112, -124),  S( 78, -155),  S(161, -159),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* blocked passed pawn */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),    // blocked by pawns
-            S(  0,   0),   S( 76, -37),   S( 55,   4),   S( 51,  24),   S( 51,  58),   S( 73,  35),   S(180,  33),   S(  0,   0),    // blocked by knights
-            S(  0,   0),   S( 33,  -1),   S( 18,  50),   S( 12,  58),   S(  7, 109),   S( 60, 116),   S(195, 134),   S(  0,   0),    // blocked by bishops
-            S(  0,   0),   S(-63,  25),   S(-17, -20),   S(-24, -23),   S(-30,   8),   S( 32,  -7),   S(271, -81),   S(  0,   0),    // blocked by rooks
-            S(  0,   0),   S( 15, -38),   S( 58, -30),   S( -1,  26),   S( -2,  -6),   S( 25, -170),  S(-49, -182),  S(  0,   0),    // blocked by queens
-            S(  0,   0),   S( -1,  16),   S( 84,   8),   S( 55, -53),   S( 92, -15),   S(235,   4),   S(390,  85),   S(  0,   0),    // blocked by kings
-
-            /* supported pawn chain */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S( 39,  -2),   S( 45,  14),   S( 61,  19),   S( 63,  32),   S( 45,  35),   S( 47,  15),   S( 36,   9),   S( 74, -12),
-            S(  0,   0),   S( 24,  23),   S( 24,  13),   S( 36,  33),   S( 36,  12),   S( 21,  16),   S( 40,   4),   S( 16,   1),
-            S(-16,  13),   S( 26,  28),   S( 64,  36),   S( 51,  34),   S( 64,  35),   S( 68,  14),   S( 22,  32),   S( 31,   6),
-            S( 70,  50),   S(151,  27),   S(137,  85),   S(171,  60),   S(186,  56),   S( 86,  82),   S(101,  24),   S(106,   0),
-            S(105,  84),   S(241,  52),   S(224, 144),   S(196,  43),   S(261, 130),   S(164, 118),   S(225,  29),   S(-64, 111),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* enemy king outside passed pawn square */
-            S( 17, 207),
-
-            /* passed pawn/friendly king distance penalty */
-            S( -2, -21),
-
-            /* passed pawn/enemy king distance bonus */
-            S(  0,  36),
-
-            /* pawn ram bonuses/penalties */
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),
-
-            /* piece threats */
-            /*  Pawn          Knight         Bishop          Rook          Queen           King */
-            S(  0,   0),   S( 67,  73),   S( 49,  92),   S( 76,  60),   S( 59,   9),   S(  0,   0),    // Pawn threats
-            S(  0,   0),   S(-12,   6),   S( 48,  28),   S( 89,  -1),   S( 29,  55),   S(  0,   0),    // Knight threats
-            S(  0,   0),   S( 31,  62),   S( -5,  37),   S( 55,  32),   S( 33, 160),   S(  0,   0),    // Bishop threats
-            S(  0,   0),   S( 21,  55),   S( 29,  54),   S(-14,  65),   S( 52,  91),   S(  0,   0),    // Rook threats
-            S(  0,   0),   S(  0,  40),   S( -5,  75),   S(-10,  29),   S(-14,  16),   S(  0,   0),    // Queen threats
-            S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),   S(  0,   0),    // King threats
-
-            /* pawn push threats */
-            S(  0,   0),   S( 32,  31),   S( 35, -22),   S( 34,  27),   S( 35, -25),   S(  0,   0),    // Pawn push threats
-
-            /* king on open diagonal */
-            S( -8,  11),
-        };
-#endif
-
-        // Solution sample size: 12000000, generated on Tue, 26 Sep 2023 12:04:40 GMT
-        // Solution error: 0.117989, accuracy: 0.5222, seed: -1214095863
-        private static readonly short[] paragonWeights =
-        {
-            /*------------------- OPENING WEIGHTS -------------------*/
-
-            /* opening piece values */
-            118, 510, 557, 696, 1589, 0,
-
-            /* opening piece square values */
-
-            #region opening piece square values
-
-            /* pawns: KK */
-               0,    0,    0,    0,    0,    0,    0,    0,
-             -57,  -55,  -37,  -17,  -47,   59,   65,   -4,
-             -60,  -48,  -50,  -62,  -36,  -12,   16,  -48,
-             -29,  -31,  -10,  -27,   -8,   23,    4,   -2,
-             -12,  -28,    1,   23,   35,   21,   44,   60,
-              56,    1,   -2,   73,  125,  196,  190,  181,
-              25,   36,    8,   46,   27,   99,   13,   40,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* pawns: KQ */
-               0,    0,    0,    0,    0,    0,    0,    0,
-             -95,  -96,  -64,  -69,  -42,   55,   68,   -6,
-             -86,  -84,  -67,  -74,  -36,  -17,   23,  -44,
-             -63,  -37,  -26,  -31,  -18,    1,    0,   10,
-             -50,  -11,    2,   17,    4,  -28,   23,   61,
-             105,   96,  140,  108,  123,   66,   71,  120,
-             106,   34,   87,   47,   57,    7,   91,   93,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* pawns: QK */
-               0,    0,    0,    0,    0,    0,    0,    0,
-             -33,    5,   -2,  -40,  -58,   -2,  -22,  -66,
-             -26,    4,  -43,  -54,  -49,  -27,  -40, -102,
-               1,    5,  -13,  -29,  -27,    3,  -20,  -42,
-              17,  -16,    3,   29,   18,    4,   12,    5,
-              71,    1,    7,   75,  132,  178,  142,   98,
-              59,  131,   55,   97,   46,   82,   21,   -1,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* pawns: QQ */
-               0,    0,    0,    0,    0,    0,    0,    0,
-             -53,   -6,   37,    8,   -6,   15,    2,  -18,
-             -47,  -21,  -19,  -45,  -13,  -34,  -28,  -71,
-             -25,  -22,   18,   -5,    9,   12,   -5,  -25,
-              56,   58,   59,   55,   27,   -7,   22,   40,
-             224,  143,  174,  146,   50,   68,   24,  121,
-             135,   33,   78,   28,   76,  -19,   97,   67,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* knights: KK */
-            -114,  -21,  -48,  -28,   -8,   -5,  -19,  -64,
-             -35,  -47,   -5,   10,   11,   21,   -4,   -5,
-             -24,   -6,    5,   23,   33,   13,    9,    3,
-              -7,   14,   26,   25,   30,   49,   54,   17,
-              15,   20,   46,   50,   19,   45,   17,   69,
-             -31,   26,   46,   54,   97,  182,   40,   -9,
-             -13,  -24,    7,   62,   48,  134,   19,  -13,
-            -309,  -53,  -30,   72,   72, -110,    4, -207,
-
-            /* knights: KQ */
-             -95,  -46,  -63,  -55,  -33,  -27,  -42,  -24,
-             -86,  -41,  -11,   15,   -4,   20,  -67,  -51,
-             -46,   -3,   -5,   27,   23,   26,   -9,  -20,
-              -6,   51,   39,   61,   52,   29,   27,  -16,
-              63,   29,   70,   69,   89,   70,   54,   21,
-              56,   75,  147,  137,   73,   75,   -4,   -4,
-               7,  -14,   10,   80,   34,   33,  -25,  -42,
-            -295,  -13,  -11,   39,  -43,   -2,  -32, -122,
-
-            /* knights: QK */
-            -101,  -75,  -52,  -74,  -30,  -25,  -57,  -93,
-             -44,  -32,   11,   -6,  -12,   -1,  -63,  -17,
-             -15,   -2,   25,   49,   61,  -18,    2,  -55,
-              29,   57,   53,   39,   51,   40,   39,  -13,
-              19,   62,   77,  127,   74,   85,   55,   75,
-              -1,   15,   45,   87,  110,  166,   47,   24,
-             -24,    4,   21,   73,  -11,   74,   -2,   64,
-            -148,  -34,  -23,   29,   -6,  -34,  -33,  -59,
-
-            /* knights: QQ */
-               9,  -38,   30,  -72,  -40,  -47,  -47,  -66,
-             -36,  -34,   15,   29,  -16,   14,    1, -139,
-              10,    9,   24,   39,   31,  -12,  -14,  -29,
-              18,   27,   78,   83,   53,   49,   52,  -34,
-              50,   25,   71,   87,   86,   73,   74,   13,
-              78,  -36,   52,  149,   64,   41,   72,    6,
-             -21,   24,   -7,   21,   54,   52,  -12,  -57,
-            -122,    0,   -1,    0,  -20,   15,    5,  -50,
-
-            /* bishops: KK */
-              19,   20,    1,  -19,    5,    3,    7,    4,
-              21,   21,   19,   12,   15,   27,   45,   30,
-              -1,   20,   31,   21,   19,   21,   23,    6,
-              -6,    1,   20,   49,   40,   19,   10,   35,
-              -3,   24,   15,   56,   37,   50,   26,    6,
-               4,   20,   -3,    2,   14,   84,   58,   -7,
-             -43,  -21,   -7,  -24,  -29,    8,  -77,  -57,
-            -100,  -61,  -36,  -37,  -48,  -83,    9,  -70,
-
-            /* bishops: KQ */
-             -50,   29,  -28,  -19,   18,    6,  102,   16,
-             -21,   -7,   -1,   10,   19,   51,   84,   19,
-             -29,    8,   31,   32,   29,   58,   59,   48,
-              22,    4,   18,   68,   85,   45,   24,   -6,
-              10,   55,   44,  136,   92,   36,   37,   27,
-               8,  110,  135,   14,   47,   43,   32,    7,
-             -50,   33,  -43,  -29,    3,   -5,  -17,  -61,
-             -32,  -34,   17,  -14,  -11,  -21,  -27,  -49,
-
-            /* bishops: QK */
-              41,   34,   22,  -24,   -8,  -28,  -29, -134,
-              60,   70,   30,   36,   -4,  -13,  -28,  -10,
-              14,   47,   51,   29,   19,    6,   11,  -16,
-               4,   34,   33,   84,   60,   14,  -24,  -10,
-             -11,   30,   48,   99,   55,   57,   50,  -14,
-             -29,    0,   47,   35,   58,   80,   92,   22,
-             -40,  -56,  -35,  -23,  -59,    3,  -24,  -29,
-             -56,  -57,  -24,  -44,    3,  -33,   26,  -27,
-
-            /* bishops: QQ */
-             -76,  -24,   19,  -61,    0,  -15,   39,  -19,
-              49,   68,   -8,   41,  -28,   39,   -6,   92,
-              12,   43,   48,   22,   39,   22,   51,   -4,
-              50,   60,   25,   87,   34,   69,   19,   25,
-              34,   37,   37,   95,   97,   12,   37,    6,
-             -36,   91,   97,   63,    5,   72,   -4,    1,
-             -47,  -24,   36,    0,   30,  -18,   37,  -18,
-             -27,   -6,   -4,   -1,   19,   -8,    7,  -30,
-
-            /* rooks: KK */
-             -33,  -21,  -16,   -3,    5,   -1,   12,  -29,
-             -49,  -32,  -25,  -15,  -13,   11,   36,  -19,
-             -50,  -47,  -34,  -21,    1,   -5,   30,   -2,
-             -44,  -31,  -22,  -10,  -19,  -26,   28,  -22,
-             -24,   -8,    2,   26,    0,   41,   73,   24,
-             -10,   28,   35,   38,   74,  141,  131,   68,
-             -31,  -30,   -3,   33,   19,   87,  102,  141,
-              56,   45,   38,   56,   36,   98,  115,  126,
-
-            /* rooks: KQ */
-             -51,   -4,  -15,  -24,  -24,  -18,  -32,  -59,
-             -70,  -17,  -31,  -29,  -26,  -38,  -42,  -42,
-             -49,  -26,  -45,  -18,  -52,  -27,   -6,  -50,
-             -69,  -14,  -65,   -7,  -19,  -50,   -7,   -8,
-             -11,   45,   -8,   -8,   44,   33,   41,    8,
-              79,   91,   56,   79,   78,   56,  130,   54,
-             -19,   17,   54,   47,   73,   13,   50,   96,
-              19,   51,   55,   46,   40,   74,   73,  110,
-
-            /* rooks: QK */
-             -93,  -27,  -32,  -24,  -16,  -16,   -4,  -45,
-             -48,  -48,  -50,  -36,  -50,  -50,  -22,  -35,
-             -39,   -4,  -33,  -43,  -66,  -72,  -22,  -32,
-              -4,   44,  -32,  -33,  -51,  -50,   10,  -23,
-              61,   34,   60,  -14,  -24,   -8,   18,    6,
-              44,  100,   40,  -12,   56,  101,  107,   47,
-              34,   45,   78,    3,  -34,   77,   34,   12,
-             242,  103,   80,   40,   11,   72,   94,   95,
-
-            /* rooks: QQ */
-             -55,  -22,   14,   17,   -4,  -17,  -33,  -44,
-             -32,  -39,  -32,    8,  -29,  -43,  -26,  -29,
-             -79,   28,    4,  -12,  -37,  -51,  -47,  -52,
-             -39,   46,  -32,  -15,  -45,  -26,  -48,  -31,
-              38,   69,    7,  110,   41,   39,    2,    7,
-              78,   92,  117,   94,   70,   67,   31,   19,
-              89,   65,   21,   78,  158,   64,   48,   13,
-              43,   32,    6,   -8,    8,   55,   97,   13,
-
-            /* queens: KK */
-              38,   46,   56,   60,   80,   59,  -11,    2,
-              26,   49,   50,   59,   61,   87,   83,   50,
-              22,   38,   45,   33,   51,   38,   50,   47,
-              33,   15,   24,   20,   20,   31,   54,   16,
-              24,   32,   -8,   -4,    5,   17,   24,   21,
-              15,   22,  -13,  -30,   11,  106,   55,   18,
-             -16,  -33,  -27,  -15,  -56,   80,   21,  126,
-             -34,   33,   16,    2,   27,  108,  159,  125,
-
-            /* queens: KQ */
-              18,   14,   27,   19,   20,    7,  -30,  -26,
-              32,   30,   44,   22,   28,   20,  -37,  -51,
-               9,    6,   40,   23,   31,   27,   12,   -5,
-               1,   57,  -20,  -14,   17,   -9,    3,  -14,
-              19,   31,    1,   -3,   41,   -4,    6,    5,
-              51,   92,   62,   25,   11,   20,   28,   16,
-              57,   95,   71,   73,   39,  -10,  -17,   21,
-              77,  125,   93,   81,   25,   56,   49,  -14,
-
-            /* queens: QK */
-             -74,  -87,   -7,   -5,   16,   -2,  -28,  -59,
-             -67,  -34,   40,   23,   21,    1,   41,    2,
-             -47,    0,    7,  -12,   14,  -18,   14,   -3,
-             -15,  -41,  -27,   -6,  -18,   -2,   11,  -13,
-              -8,  -12,  -43,  -37,  -31,   40,    5,   13,
-              -2,   -4,  -18,  -30,   11,   69,   40,   18,
-               6,    0,  -53,  -43,  -26,   -3,   49,   61,
-             -37,   12,   25,   22,  -34,   76,   92,   61,
-
-            /* queens: QQ */
-             -58,  -73,  -39,  -40,   10,  -13,  -67,  -76,
-             -77,  -49,    8,   24,  -36,  -23,  -16,  -61,
-             -23,  -13,   10,  -11,  -13,  -31,  -19,  -38,
-             -45,   26,  -38,    1,   -5,   -1,  -17,  -38,
-             -27,  -19,   45,   37,   66,    5,  -16,   -1,
-              36,   45,   90,   47,    3,  -32,  -12,   -9,
-              82,   51,   66,   26,    6,  -22,  -11,  -45,
-               8,   69,   16,   24,   12,   11,   26,   12,
-
-            /* kings: KK */
-               0,    0,    0,    0, -108, -125,  -50,  -60,
-               0,    0,    0,    0,  -57,  -47,  -27,  -53,
-               0,    0,    0,    0,   -8,  -46,  -50, -103,
-               0,    0,    0,    0,  109,   59,    9, -128,
-               0,    0,    0,    0,  228,  284,  175,   31,
-               0,    0,    0,    0,  250,  295,  266,  150,
-               0,    0,    0,    0,  204,  149,  146,   63,
-               0,    0,    0,    0,   64,   68,   66,   -7,
-
-            /* kings: KQ */
-               0,    0,    0,    0, -146, -162,  -68,  -78,
-               0,    0,    0,    0,  -51,  -54,  -59,  -81,
-               0,    0,    0,    0,  -42,  -15,  -39, -114,
-               0,    0,    0,    0,  128,   56,  -17,  -84,
-               0,    0,    0,    0,  227,  142,   81,   -3,
-               0,    0,    0,    0,  174,   94,  110,   -7,
-               0,    0,    0,    0,  180,  217,   99,   30,
-               0,    0,    0,    0,   71,   73,   81,    2,
-
-            /* kings: QK */
-             -64,  -41,  -86, -160,    0,    0,    0,    0,
-             -37,  -58,  -69,  -86,    0,    0,    0,    0,
-             -97,  -69,  -38,  -80,    0,    0,    0,    0,
-             -98,    2,   47,   42,    0,    0,    0,    0,
-             -48,   60,   78,  158,    0,    0,    0,    0,
-             -31,   81,  153,  167,    0,    0,    0,    0,
-              43,  141,  118,  137,    0,    0,    0,    0,
-             -28,   38,  124,   51,    0,    0,    0,    0,
-
-            /* kings: QQ */
-            -177, -111, -157, -201,    0,    0,    0,    0,
-            -121, -114,  -84, -121,    0,    0,    0,    0,
-            -129,  -73,  -59,  -58,    0,    0,    0,    0,
-             -61,   15,   42,  106,    0,    0,    0,    0,
-              51,  147,  214,  214,    0,    0,    0,    0,
-              63,  169,  163,  123,    0,    0,    0,    0,
-               5,   37,   73,   95,    0,    0,    0,    0,
-             -24,   25,   34,   64,    0,    0,    0,    0,
-
-            #endregion
-
-            /* opening mobility weights */
-
-               9, // knights
-               5, // bishops
-               3, // rooks
-               1, // queens
-
-            /* opening squares attacked near enemy king */
-              24, // attacks to squares 1 from king
-              21, // attacks to squares 2 from king
-               7, // attacks to squares 3 from king
-
-            /* opening pawn shield/king safety */
-              21, // # friendly pawns 1 from king
-               8, // # friendly pawns 2 from king
-               6, // # friendly pawns 3 from king
-
-            /* opening isolated pawns */
-            -4,
-
-            /* opening unused */
-            0,
-
-            /* opening doubled pawns */
-            -16,
-
-            /* opening adjacent/connected pawns */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              28,   -1,   10,   15,   31,    1,  -30,    9,
-              18,   23,   17,   21,   57,    0,   19,   13,
-             -10,   23,    0,   29,   39,   40,   23,   20,
-              42,   14,   70,   14,  107,  107,   13,   38,
-              72,  143,  180,  191,  172,  169,  155,   84,
-              77,  119,  116,  118,   81,   52,   45,   26,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* opening knight on outpost */
-            -1,
-
-            /* opening bishop on outpost */
-            6,
-
-            /* opening bishop pair */
-            40,
-
-            /* opening rook on open file */
-            35,
-
-            /* opening rook on half-open file */
-            12,
-
-            /* opening rook behind passed pawn */
-            0,
-
-            /* opening doubled rooks on file */
-            9,
-
-            /* opening king on open file */
-            -79,
-
-            /* opening king on half-open file */
-            -32,
-
-            /* opening castling rights available */
-            38,
-
-            /* opening castling complete */
-            19,
-
-            /* opening center control */
-               0, // D0
-               3, // D1
-
-            /* opening queen on open file */
-            -10,
-
-            /* opening queen on half-open file */
-            7,
-
-            /* opening rook on seventh rank */
-            15,
-
-            /* opening passed pawn */
-               0,    0,    0,    0,    0,    0,    0,    0,
-               2,  -12,  -25,  -47,   17,  -26,  -22,   30,
-              12,   -1,  -16,   -9,  -33,  -51,  -60,   20,
-              40,   24,    0,   30,    0,  -27,  -54,   -1,
-              47,   68,   25,    5,    7,   39,   28,   -7,
-              52,   90,   86,   10,  -33,   61,   13,  -21,
-             153,  136,  120,  129,  134,  124,  177,  153,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* opening bad bishop pawns */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              -2,    1,   -4,   -3,  -12,  -29,  -23,    0,
-              -5,   -9,  -15,   -8,  -15,  -17,  -16,   -2,
-              -6,    3,    0,   -2,  -18,  -13,  -12,   -2,
-               4,    8,   -1,    2,   -7,   -8,   -6,   -3,
-              21,    8,   17,    8,   32,   -6,   49,  -11,
-              59,   98,   94,   88,   82,   94,   66,  150,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* opening block passed pawn */
-               0,    0,    0,    0,    0,    0,    0,    0,  // blocked by pawns
-               0,   76,   55,   51,   51,   73,  200,    0,  // blocked by knights
-               0,   33,   18,   12,    7,   60,  214,    0,  // blocked by bishops
-               0,  -63,  -17,  -25,  -30,   31,  288,    0,  // blocked by rooks
-               0,   15,   58,   -1,   -2,   26,  -33,    0,  // blocked by queens
-               0,    0,   84,   59,   99,  239,  421,    0,  // blocked by kings
-
-            /* opening supported pawn chain */
-               0,    0,    0,    0,    0,    0,    0,    0,
-               0,    0,    0,    0,    0,    0,    0,    0,
-              36,   37,   51,   53,   34,   36,   24,   68,
-              -6,   21,   21,   30,   35,   19,   40,   10,
-             -20,   29,   64,   49,   65,   72,   28,   29,
-              68,  153,  138,  171,  187,   89,  104,  103,
-             101,  238,  220,  192,  256,  159,  220,  -68,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* opening enemy king outside passed pawn square */
-            16,
-
-            /* opening passed pawn/friendly king distance penalty */
-            -6,
-
-            /* opening passed pawn/enemy king distance bonus */
-            -2,
-
-            /* opening pawn rams */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              48,   19,   20,   23,   95,   29,   26,   61,
-              10,    5,   24,   26,   16,   -5,   12,   30,
-               0,    0,    0,    0,    0,    0,    0,    0,
-             -10,   -5,  -24,  -26,  -16,    5,  -12,  -30,
-             -48,  -19,  -20,  -23,  -95,  -29,  -26,  -61,
-               0,    0,    0,    0,    0,    0,    0,    0,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* opening piece threats */
-            /* P     N     B     R     Q     K */
-               0,   67,   50,   76,   59,    0,  // Pawn threats
-               0,  -12,   48,   89,   29,    0,  // Knight threats
-               0,   31,   -5,   55,   33,    0,  // Bishop threats
-               0,   22,   29,  -14,   52,    0,  // Rook threats
-               0,    0,   -5,  -10,  -14,    0,  // Queen threats
-               0,    0,    0,    0,    0,    0,  // King threats
-
-            /* opening pawn push threats */
-               0,   32,   35,   34,   35,    0,  // Pawn push threats
-
-            /* opening king on open diagonal */
-            -8,
-
-            /* opening passed pawn can advance */
-              12,   38,   53,  249,
-
-
-            /*------------------- END GAME WEIGHTS -------------------*/
-
-            /* end game piece values */
-            148, 451, 500, 866, 1527, 0,
-
-            /* end game piece square values */
-
-            #region end game piece square values
-
-            /* pawns: KK */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              17,   17,    0,  -23,   33,   -7,  -29,  -35,
-              11,   -3,  -22,  -14,  -21,  -18,  -24,  -15,
-              15,   10,  -29,  -44,  -31,  -31,   -9,  -23,
-              63,   35,   16,   -2,  -26,  -16,    0,   -8,
-              73,   88,   81,   62,   65,   15,   28,   16,
-              58,   57,   70,   23,   34,  -39,   23,  -25,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* pawns: KQ */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              34,   43,   19,   52,  -16,  -36,  -41,  -15,
-               9,    9,  -20,  -25,  -44,  -32,  -37,   -7,
-              28,   10,  -14,  -34,  -30,  -26,  -20,  -23,
-              76,    9,   -1,   11,   -7,   32,   19,   10,
-              41,    2,  -22,   38,  101,  120,  121,  103,
-             -30,   -1,  -43,   39,   73,  123,   86,   76,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* pawns: QK */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              20,    4,   -5,   21,   17,   -3,    0,   22,
-              16,  -10,  -14,  -25,  -20,  -25,  -19,    9,
-              14,   10,   -9,  -30,  -16,  -28,  -17,   -7,
-              69,   61,   42,   18,  -12,  -18,   -8,   15,
-              94,  133,  130,   89,   60,    3,   28,   51,
-              46,   64,  117,   59,   23,  -38,  -12,   -4,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* pawns: QQ */
-               0,    0,    0,    0,    0,    0,    0,    0,
-               6,   -1,  -23,  -14,  -18,  -35,  -33,  -21,
-              -2,   -7,  -30,  -38,  -47,  -48,  -43,  -17,
-              11,    5,  -34,  -42,  -46,  -56,  -35,  -28,
-              20,   -4,  -18,   -2,  -28,  -17,  -21,   -9,
-             -24,  -10,  -36,   15,   78,   52,   63,   49,
-             -48,   10,  -52,   23,   -2,   46,   18,   31,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* knights: KK */
-             -40,  -76,    1,   16,    8,   10,  -10,  -57,
-             -41,   28,    7,   11,   21,   33,   26,    7,
-              -9,   23,   35,   62,   55,   28,   39,   34,
-              34,   54,   80,   91,   99,   81,   73,   45,
-              48,   48,   61,   91,  101,  107,  114,   54,
-              67,   50,   59,   60,   45,   19,   75,   66,
-              40,   62,   64,   61,   67,    1,   24,   45,
-              59,   69,   69,   47,   44,   50,   47,  -74,
-
-            /* knights: KQ */
-            -110,  -48,    2,   13,  -12,  -15,   15, -128,
-             -15,   11,  -21,  -17,   -5,    1,   32,  -10,
-              -7,  -30,    6,    3,    2,  -32,   14,   12,
-               3,  -12,   19,    8,   18,   26,   15,   23,
-             -46,  -20,   -3,    5,   -2,    2,   -5,   -5,
-             -39,  -64,  -70,  -51,   -5,    0,   31,    1,
-             -45,  -29,  -43,  -17,   -7,    7,   22,    1,
-            -119,  -55,  -35,   -8,  -30,   36,   34,  -54,
-
-            /* knights: QK */
-            -104,   24,   21,   45,   -4,  -18,  -35,  -80,
-             -61,   16,   -4,   10,    8,    7,    5,  -57,
-               9,  -13,  -10,   21,   -2,   16,  -14,  -27,
-               0,   11,   23,   49,   22,   28,    5,    2,
-              25,   12,   23,   -1,   25,   15,    4,  -22,
-              35,   31,   31,   -9,   -6,  -39,   -9,  -25,
-              26,   13,   38,   42,   34,  -10,  -20,  -35,
-             -83,   23,   36,   -4,   25,  -26,  -26,  -99,
-
-            /* knights: QQ */
-            -108,  -15,  -29,    9,   -2,    3,  -72,  -82,
-             -41,   28,  -31,  -20,  -10,  -53,  -18,  -39,
-             -23,  -24,  -26,   -6,   -6,  -17,   -3,  -48,
-              43,    7,    1,  -15,   -6,  -14,    0,   11,
-              -9,    9,  -17,  -15,   -8,  -21,  -29,    2,
-             -24,   -1,  -27,  -37,  -22,    9,  -26,  -26,
-              30,  -12,   -1,   25,   11,    2,   10,   16,
-             -85,   58,    6,   30,  -15,   33,   48,  -45,
-
-            /* bishops: KK */
-             -15,  -17,    3,   46,   37,   42,   11,  -49,
-               2,   10,   22,   44,   39,   32,    7,  -45,
-              26,   42,   62,   52,   76,   47,   21,   21,
-              48,   63,   72,   56,   46,   51,   35,   -9,
-              66,   36,   50,   48,   55,   49,   45,   59,
-              50,   58,   58,   57,   65,   62,   48,   88,
-              75,   65,   67,   76,   74,   54,   61,   50,
-              95,   86,   83,   89,   83,   71,   28,   56,
-
-            /* bishops: KQ */
-               2,  -10,   13,    5,   -9,  -15,  -92,  -95,
-             -16,   -2,   12,    8,    3,  -25,  -27,  -71,
-               7,    8,   13,    4,   15,  -10,  -16,  -37,
-              15,   22,    3,  -11,   -6,   24,   17,   19,
-              22,  -12,  -13,  -62,   -1,   22,   18,    5,
-             -10,  -37,  -30,   14,    5,   11,   23,   35,
-             -31,  -17,   22,   18,   10,   21,   44,   68,
-               5,    6,   -5,   21,   19,   35,   38,   36,
-
-            /* bishops: QK */
-             -65,  -35,    3,   27,   18,    1,   27,    0,
-             -57,  -23,   12,   15,   15,   14,   -3,  -19,
-              -2,    8,   15,   27,   35,    4,   -7,  -15,
-               2,   20,   41,    3,   -5,   28,   30,   19,
-              47,   21,   17,   -4,    7,   13,  -13,   31,
-              46,   47,   16,   32,   25,   20,  -29,    8,
-              53,   44,   51,   50,   49,   29,   -4,  -27,
-              35,   40,   42,   55,   14,    2,    7,   18,
-
-            /* bishops: QQ */
-             -20,  -18,  -32,   -8,  -20,  -34,  -51,  -82,
-             -74,  -49,  -10,  -25,    4,  -46,  -34, -107,
-             -21,  -27,  -24,  -20,  -17,  -27,  -57,  -23,
-             -19,  -21,  -22,  -43,  -29,  -26,  -23,  -32,
-             -16,  -21,  -34,  -48,  -57,  -22,  -36,   -7,
-              -3,  -24,  -37,  -25,    3,  -38,   -7,   -6,
-             -23,  -14,  -24,    5,  -23,  -11,  -19,    3,
-             -16,  -11,  -20,    8,    9,   -4,    9,   34,
-
-            /* rooks: KK */
-             101,   93,   80,   53,   38,   65,   61,   52,
-              98,   89,   77,   52,   40,   24,   39,   53,
-             101,  103,   77,   52,   30,   44,   34,   44,
-             123,  109,   93,   72,   55,   86,   71,   75,
-             134,  118,   98,   67,   84,   70,   74,   82,
-             133,  111,   94,   66,   48,   52,   65,   74,
-             129,  130,   94,   63,   69,   65,   88,   62,
-             105,  121,   94,   62,   94,   96,  108,   81,
-
-            /* rooks: KQ */
-              15,  -16,  -26,  -11,  -10,   -7,   35,   42,
-              13,  -32,  -30,  -27,  -19,    3,   22,   11,
-               2,  -16,  -19,  -39,  -13,   -9,    1,    7,
-              16,  -10,    4,  -20,  -18,   26,   17,    6,
-               7,  -15,   -4,    5,  -26,   -3,   12,   24,
-             -11,  -20,  -35,  -42,  -42,   -7,  -20,   14,
-              21,    9,  -32,  -25,  -28,   -5,   14,   -7,
-              39,   14,  -13,  -14,  -10,   -2,   30,    0,
-
-            /* rooks: QK */
-              55,   26,   22,  -11,  -25,  -25,  -25,   12,
-              20,   17,   23,  -12,  -19,  -23,  -28,  -15,
-              17,   16,    8,   -9,   -1,   -7,  -32,  -30,
-              10,   -2,   21,    8,    9,   11,   -5,   -3,
-              11,   33,    0,   20,   12,    7,    3,  -12,
-              26,   11,   23,   17,  -17,  -12,  -13,    2,
-              18,   20,   -2,   11,   20,  -10,   14,   17,
-            -105,   -1,    3,   14,   19,   -3,    6,   -3,
-
-            /* rooks: QQ */
-             -14,  -32,  -69,  -87,  -63,  -30,   -5,   15,
-             -47,  -26,  -45,  -86,  -74,  -57,  -44,  -16,
-             -14,  -54,  -65,  -72,  -36,  -49,  -12,   -4,
-             -13,  -47,  -27,  -45,  -48,  -32,   -8,    4,
-             -31,  -42,  -30,  -79,  -50,  -34,   -1,    0,
-             -37,  -29,  -49,  -63,  -63,  -31,  -10,    0,
-             -39,  -24,  -11,  -57,  -94,  -58,  -22,    0,
-               3,   12,   -4,   -4,  -38,  -33,   -2,    9,
-
-            /* queens: KK */
-              56,   29,    3,   14,  -50,  -76,  -25,   56,
-              84,   53,   49,   31,   21,  -32,  -50,  -16,
-              61,   63,   71,   70,   53,   85,   79,   30,
-              82,  105,   89,   96,   87,  105,   87,  139,
-              95,  100,  137,  127,  140,  170,  184,  197,
-             111,  146,  150,  161,  164,  155,  206,  252,
-             165,  195,  183,  156,  206,  140,  278,  167,
-             166,  128,  135,  154,  151,  144,  154,  173,
-
-            /* queens: KQ */
-             -25,  -13,  -91,  -48,  -25,  -59,  -26,   15,
-               8,  -34,  -85,  -34,  -37,  -60,  -27,  -47,
-               2,  -20,  -86,  -33,  -29,   19,  -11,  -57,
-              32,  -78,   10,   -9,  -11,   29,   30,   51,
-             -10,  -13,   -9,  -18,  -35,   27,   38,   68,
-               0,    4,   -1,   -5,   17,   37,   69,   47,
-              27,   10,   17,    4,   -3,   70,   89,   52,
-              40,   36,   34,   53,   67,   28,   24,   24,
-
-            /* queens: QK */
-             -55,  -30, -105,  -69, -103,  -61,  -27,   32,
-             -24,  -85, -117,  -71,  -72,  -52,  -64,    1,
-             -33,  -23,  -11,   11,  -34,   -5,  -49,    0,
-             -31,   33,   41,    3,   12,  -17,  -19,   -5,
-              27,   10,   66,   44,   27,  -18,   21,  -22,
-              17,   72,   44,   46,   23,    9,   31,   26,
-              -8,   32,   85,   43,   27,   27,   11,   35,
-              43,   35,   26,   23,   29,   41,   31,   45,
-
-            /* queens: QQ */
-             -52,  -45,  -14,  -57,  -36,  -59,  -30,  -13,
-             -30,  -53,  -69,  -89,  -20,  -72,  -47,  -35,
-             -18,   -9,  -59,  -54,  -60,  -35,  -49,  -37,
-             -24,  -23,   -4,  -54,  -63, -103,  -59,    2,
-               3,   13,  -43,  -48,  -50,  -44,  -11,  -44,
-              13,   14,   39,  -27,  -25,  -13,   -2,   50,
-              23,   42,   12,   32,   -6,  -14,   20,    7,
-             -34,   42,   26,   11,  -10,  -30,  -23,  -29,
-
-            /* kings: KK */
-               0,    0,    0,    0,   -2,   23,   -7,  -38,
-               0,    0,    0,    0,   10,   10,    6,    6,
-               0,    0,    0,    0,   16,   20,   12,   12,
-               0,    0,    0,    0,    5,   18,   16,   24,
-               0,    0,    0,    0,  -20,  -36,   -6,   -2,
-               0,    0,    0,    0,  -19,  -29,    3,    4,
-               0,    0,    0,    0,  -14,   -4,   51,   14,
-               0,    0,    0,    0,    2,   31,   38, -139,
-
-            /* kings: KQ */
-               0,    0,    0,    0,   18,   25,  -15,  -25,
-               0,    0,    0,    0,   -1,    1,   18,   12,
-               0,    0,    0,    0,   15,    0,    9,   19,
-               0,    0,    0,    0,   -5,   20,   35,   33,
-               0,    0,    0,    0,  -11,   22,   36,   39,
-               0,    0,    0,    0,    6,   42,   38,   57,
-               0,    0,    0,    0,   -2,   -9,   27,   23,
-               0,    0,    0,    0,   -7,    2,  -22,  -22,
-
-            /* kings: QK */
-             -71,  -68,  -45,  -16,    0,    0,    0,    0,
-             -47,  -15,  -30,  -23,    0,    0,    0,    0,
-             -17,   -7,  -21,   -3,    0,    0,    0,    0,
-              13,    5,   -6,   -8,    0,    0,    0,    0,
-              12,   20,    7,  -20,    0,    0,    0,    0,
-              34,   26,    4,  -12,    0,    0,    0,    0,
-              -3,    4,    3,   -7,    0,    0,    0,    0,
-             -21,   21,   -3,    1,    0,    0,    0,    0,
-
-            /* kings: QQ */
-              46,    4,   14,    3,    0,    0,    0,    0,
-              12,   13,   -9,  -10,    0,    0,    0,    0,
-               4,   -8,  -11,   -7,    0,    0,    0,    0,
-             -13,    2,    2,  -13,    0,    0,    0,    0,
-             -10,   -3,  -17,  -22,    0,    0,    0,    0,
-              15,   18,  -17,   10,    0,    0,    0,    0,
-              -4,   64,   15,    7,    0,    0,    0,    0,
-            -134,   51,   10,   10,    0,    0,    0,    0,
-
-            #endregion
-
-            /* end game mobility weights */
-
-               7, // knights
-               4, // bishops
-               3, // rooks
-               3, // queens
-
-            /* end game squares attacked near enemy king */
-              -5, // attacks to squares 1 from king
-              -2, // attacks to squares 2 from king
-               1, // attacks to squares 3 from king
-
-            /* end game pawn shield/king safety */
-              16, // # friendly pawns 1 from king
-              19, // # friendly pawns 2 from king
-              14, // # friendly pawns 3 from king
-
-            /* end game isolated pawns */
-            -11,
-
-            /* end game unused */
-            0,
-
-            /* end game doubled pawns */
-            -33,
-
-            /* end game adjacent/connected pawns */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              -8,   -1,   -2,   33,   31,  -43,   53,  -49,
-               1,   -9,   27,   41,   -5,   22,    2,  -19,
-              10,   19,   50,   79,   28,   37,    6,   21,
-              67,   52,   88,   85,   71,   36,   59,   44,
-             110,  106,  170,  165,  119,  170,   75,   26,
-             199,  289,  251,  237,  197,  158,  212,  167,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* end game knight on outpost */
-            26,
-
-            /* end game bishop on outpost */
-            25,
-
-            /* end game bishop pair */
-            103,
-
-            /* end game rook on open file */
-            3,
-
-            /* end game rook on half-open file */
-            34,
-
-            /* end game rook behind passed pawn */
-            43,
-
-            /* end game doubled rooks on file */
-            15,
-
-            /* end game king on open file */
-            15,
-
-            /* end game king on half-open file */
-            29,
-
-            /* end game castling rights available */
-            -41,
-
-            /* end game castling complete */
-            -18,
-
-            /* end game center control */
-               9, // D0
-               5, // D1
-
-            /* end game queen on open file */
-            38,
-
-            /* end game queen on half-open file */
-            31,
-
-            /* end game rook on seventh rank */
-            30,
-
-            /* end game passed pawn */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              18,   32,   31,   50,    1,   29,   59,   16,
-              24,   42,   39,   34,   45,   49,   76,   27,
-             -21,   -1,   10,    1,    3,   21,   48,   18,
-              -8,   14,   12,    5,   34,   22,   35,   37,
-              39,   57,   37,   27,   15,   11,   44,   28,
-              32,   68,   80,   70,   73,   75,   81,   53,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* end game bad bishop pawns */
-               0,    0,    0,    0,    0,    0,    0,    0,
-              -1,   -8,   -7,  -53,  -17,    4,   -6,   -6,
-             -10,  -11,  -14,  -21,  -19,   -6,   -8,  -11,
-             -12,  -27,  -41,  -56,  -35,  -30,  -23,   -7,
-             -28,  -38,  -36,  -50,  -38,  -31,  -33,  -23,
-             -36,  -53,  -62,  -63,  -78,  -53,  -96,  -19,
-             -56, -105, -110, -132, -162, -118, -152, -156,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* end game block passed pawn */
-               0,    0,    0,    0,    0,    0,    0,    0,  // blocked by pawns
-               0,  -37,    3,   24,   59,   36,   27,    0,  // blocked by knights
-               0,   -2,   50,   58,  109,  116,  128,    0,  // blocked by bishops
-               0,   25,  -20,  -22,    8,   -6,  -86,    0,  // blocked by rooks
-               0,  -38,  -30,   27,   -5, -171, -185,    0,  // blocked by queens
-               0,   15,    7,  -55,  -18,    3,   75,    0,  // blocked by kings
-
-            /* end game supported pawn chain */
-               0,    0,    0,    0,    0,    0,    0,    0,
-               0,    0,    0,    0,    0,    0,    0,    0,
-              -1,   17,   23,   34,   40,   19,   13,  -10,
-               2,   23,   13,   34,   12,   16,    4,    3,
-              14,   26,   36,   34,   34,   13,   30,    7,
-              52,   26,   85,   60,   55,   81,   23,    1,
-              85,   54,  144,   46,  132,  120,   31,  113,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* end game enemy king outside passed pawn square */
-            207,
-
-            /* end game passed pawn/friendly king distance penalty */
-            -19,
-
-            /* end game passed pawn/enemy king distance bonus */
-            37,
-
-            /* end game pawn rams */
-               0,    0,    0,    0,    0,    0,    0,    0,
-               7,   22,   25,   49,   26,   20,   15,    8,
-              14,   14,   14,   28,   12,    7,    4,   -5,
-               0,    0,    0,    0,    0,    0,    0,    0,
-             -14,  -14,  -14,  -28,  -12,   -7,   -4,    5,
-              -7,  -22,  -25,  -49,  -26,  -20,  -15,   -8,
-               0,    0,    0,    0,    0,    0,    0,    0,
-               0,    0,    0,    0,    0,    0,    0,    0,
-
-            /* end game piece threats */
-            /* P     N     B     R     Q     K */
-               0,   73,   92,   60,    9,    0,  // Pawn threats
-               0,    6,   28,   -1,   55,    0,  // Knight threats
-               0,   62,   38,   32,  159,    0,  // Bishop threats
-               0,   55,   54,   64,   91,    0,  // Rook threats
-               0,   40,   75,   29,   15,    0,  // Queen threats
-               0,    0,    0,    0,    0,    0,  // King threats
-
-            /* end game pawn push threats */
-               0,   31,  -22,   27,  -24,    0,  // Pawn push threats
-
-            /* end game king on open diagonal */
-            12,
-
-            /* end game passed pawn can advance */
-              12,   26,   41,   85,
-        };
-    }
-}
